test: Extracting SaInvoiceSelect tests
